Modularise event-store and clean unnused code
